Fix init_py file removal errors when files do not exist

+++ setup.py
@@ -419,7 +419,8 @@ class RemoveGenerated(clean.clean):
       os.remove(MP_THIRD_PARTY_BUILD)
       shutil.move(_get_backup_file(MP_THIRD_PARTY_BUILD), MP_THIRD_PARTY_BUILD)
     for init_py in SUBDIR_INIT_PY_FILES:
-      os.remove(init_py)
+      if os.path.exists(init_py):
+        os.remove(init_py)
     clean.clean.run(self)
